fix: github comments
